20170523 - Add location query to report view.   - query all locations in Location model.   - create list of all location_guid in database   - select location_guid based on selection passed by the form.
Current changes produces "TypeError: list indices must be integers or
slices, not str"20170523 - Add location query to report view.
  - query all locations in Location model.
  - create list of all location_guid in database
  - select location_guid based on selection passed by the form.

Current changes produces "TypeError: list indices must be integers or
slices, not str"

	File "/Users/jlawton/DevOps/perc/perc/main/views.py", line 68, in report
	  loc_guid = locs[location]
It appears that the value passed in from the form is represented as a
string despite being represented as integer value in the browser. Will
attempt to force integer value.